+++ tests/phpunit/unit/Framework/Kernel/PluginTest.php
@@ -27,7 +27,7 @@ class PluginTest extends TestCase {
 	 * Test plugin version constant.
 	 */
 	public function test_plugin_version_constant(): void {
-		$this->assertSame( '1.1.1', Plugin::VERSION );
+		$this->assertSame( '1.1.2', Plugin::VERSION );
 	}
 
 	/**